requirements.txt
mysql-connector-python==8.0.32
tqdm==4.65.0